refactor: for sharing options
