+++ client/src/components/SharingOptions/index.jsx
@@ -1,10 +1,10 @@
-import { Typography, Form, Input, Select, Checkbox, Button, message } from 'antd';
+import { Typography, message } from 'antd';
+import SharingOptionsForm from 'components/SharingOptionsForm';
 import LangValContext from 'contexts/langValContext';
 import { useContext, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import styled from 'styled-components';
 import { http } from 'utils';
-import { expirationList } from './config';
 
 const OptionsWrapper = styled.div`
   height: 100%;
@@ -18,20 +18,9 @@ const OptionsWrapper = styled.div`
     justify-content: center;
   }
 `;
-const layout = {
-  labelCol: { span: 8 },
-  wrapperCol: { span: 16 },
-};
-const tailLayout = {
-  wrapperCol: { offset: 8, span: 16 },
-};
-const centerLayout = {
-  wrapperCol: { offset: 6, span: 18 },
-};
 
 const Options = () => {
   const history = useHistory();
-
   const [checked, setChecked] = useState(false);
   const { codes, language } = useContext(LangValContext);
 
@@ -65,52 +54,7 @@ const Options = () => {
   return (
     <OptionsWrapper>
       <Typography.Title level={3}>Sharing Settings</Typography.Title>
-      <Form
-        {...layout}
-        onFinish={onFinish}
-        initialValues={{ expireAfterSeconds: -1 }}
-        name="options"
-        labelAlign="left"
-      >
-        <Form.Item label="Expiration Time" name="expireAfterSeconds">
-          <Select>
-            {expirationList.map(exp => (
-              <Select.Option key={exp.name} value={exp.ms}>
-                {exp.name}
-              </Select.Option>
-            ))}
-          </Select>
-        </Form.Item>
-        <Form.Item label="Password">
-          <Checkbox onChange={() => setChecked(!checked)} checked={checked}>
-            {checked ? 'Disabled' : 'Enabled'}
-          </Checkbox>
-        </Form.Item>
-        {checked && (
-          <Form.Item
-            {...tailLayout}
-            name="password"
-            rules={[
-              {
-                required: true,
-                type: 'string',
-                max: 10,
-                min: 4,
-              },
-            ]}
-          >
-            <Input.Password />
-          </Form.Item>
-        )}
-        <Form.Item label="Title/Name" name="title">
-          <Input />
-        </Form.Item>
-        <Form.Item {...centerLayout}>
-          <Button type="primary" htmlType="submit">
-            Create New Paste
-          </Button>
-        </Form.Item>
-      </Form>
+      <SharingOptionsForm onFinish={onFinish} checked={checked} setChecked={setChecked} />
     </OptionsWrapper>
   );
 };
